requirements.txt
python-telegram-bot
openpyxl
python-docx
Pillow
aiosqlite
nest_asyncio
matplotlib